Update README.md
Figured it was easier to just make the changes I was going to suggest and submit them to you as a PR. 

There are some pictures that I think we should include, so I put a description of them between asterisks (e.g. ***image description***). If you agree with the suggestion, would you be able to help me create the screenshot?